+++ src/representative_subset_calculator/streaming/bucket.h
@@ -6,29 +6,29 @@ class ThresholdBucket
     private:
     std::unique_ptr<MutableSubset> solution;
     std::unique_ptr<std::vector<const DataRow *>> solutionRows; 
-    std::unique_ptr<std::vector<double>> d; 
+    std::unique_ptr<std::vector<float>> d; 
     std::unique_ptr<std::vector<std::unique_ptr<DenseDataRow>>> b; 
 
-    double threshold;
+    float threshold;
     int k;
 
     public:
-    ThresholdBucket(const double threshold, const int k) 
+    ThresholdBucket(const float threshold, const int k) 
     : 
         threshold(threshold), 
         k(k), 
         solution(NaiveMutableSubset::makeNew()), 
         solutionRows(std::make_unique<std::vector<const DataRow *>>()),
-        d(std::make_unique<std::vector<double>>()),
+        d(std::make_unique<std::vector<float>>()),
         b(std::make_unique<std::vector<std::unique_ptr<DenseDataRow>>>())
     {}
 
     ThresholdBucket(
-        const double threshold, 
+        const float threshold, 
         const int k, 
         std::unique_ptr<MutableSubset> nextSolution,
         std::unique_ptr<std::vector<const DataRow *>> solutionRows,
-        std::unique_ptr<std::vector<double>> d,
+        std::unique_ptr<std::vector<float>> d,
         std::unique_ptr<std::vector<std::unique_ptr<DenseDataRow>>> b
     ) : 
         threshold(threshold), 
@@ -39,7 +39,7 @@ class ThresholdBucket
         b(move(b))
     {}
 
-    std::unique_ptr<ThresholdBucket> transferContents(const double newThreshold) {
+    std::unique_ptr<ThresholdBucket> transferContents(const float newThreshold) {
         return std::unique_ptr<ThresholdBucket>(
             new ThresholdBucket(
                 newThreshold, 
@@ -69,19 +69,19 @@ class ThresholdBucket
             return false;
         }
         
-        double d_i = std::sqrt(data.dotProduct(data));
+        float d_i = std::sqrt(data.dotProduct(data));
         std::unique_ptr<DenseDataRow> c_i(new DenseDataRow());
 
         for (size_t j = 0; j < this->solution->size(); j++) {
             if (!this->passesThreshold(std::log(std::pow(d_i, 2)))) {
                 return false;
             }
-            const double e_i = (data.dotProduct(*(solutionRows->at(j))) - c_i->dotProduct(*b->at(j))) / d->at(j);
+            const float e_i = (data.dotProduct(*(solutionRows->at(j))) - c_i->dotProduct(*b->at(j))) / d->at(j);
             c_i->push_back(e_i);
             d_i = std::sqrt(std::pow(d_i, 2) - std::pow(e_i, 2));
         }
 
-        const double marginal = std::log(std::pow(d_i, 2));
+        const float marginal = std::log(std::pow(d_i, 2));
         if (this->passesThreshold(marginal)) {
             this->solution->addRow(rowIndex, marginal);
             this->solutionRows->push_back(&data);
@@ -94,7 +94,7 @@ class ThresholdBucket
     }
 
     private:
-    bool passesThreshold(double marginalGain) {
+    bool passesThreshold(float marginalGain) {
         return marginalGain >= (((this->threshold / 2) - this->solution->getScore()) / (this->k - this->solution->size()));
     }
 };
