+++ src/representative_subset_calculator/streaming/bucket_titrator.h
@@ -6,13 +6,13 @@ class BucketTitrator {
     virtual void processQueueDynamicBuckets(SynchronousQueue<std::unique_ptr<CandidateSeed>> &seedQueue) = 0;
     virtual std::unique_ptr<Subset> getBestSolutionDestroyTitrator() = 0;
     virtual bool bucketsInitialized() const = 0;
-    virtual void initBuckets(const double deltaZero) = 0;
+    virtual void initBuckets(const float deltaZero) = 0;
 
     protected:
-    static size_t getNumberOfBuckets(const unsigned int k, const double epsilon) {
-        size_t numBuckets = (int)(0.5 + [](double val, double base) {
+    static size_t getNumberOfBuckets(const unsigned int k, const float epsilon) {
+        size_t numBuckets = (int)(0.5 + [](float val, float base) {
             return log2(val) / log2(base);
-        }((double)k, (1 + epsilon))) + 1;
+        }((float)k, (1 + epsilon))) + 1;
 
         return numBuckets;
     }
@@ -21,20 +21,20 @@ class BucketTitrator {
 class ThreeSieveBucketTitrator : public BucketTitrator {
     private:
     bool firstBucketBuilt;
-    const double epsilon;
+    const float epsilon;
     const unsigned int T; // hyperparameter -- tranfer contents to next bucket after T unsuccesful inserts
     const unsigned int k;
 
     unsigned int t; // counts till T
     size_t currentBucketIndex;
     size_t totalBuckets;
-    double deltaZero;
+    float deltaZero;
     std::unique_ptr<ThresholdBucket> bucket;
     std::vector<std::unique_ptr<CandidateSeed>> seedStorage;
 
     public:
     ThreeSieveBucketTitrator(
-        const double epsilon,
+        const float epsilon,
         const unsigned int T,
         const unsigned int k
     ) : 
@@ -106,9 +106,9 @@ class ThreeSieveBucketTitrator : public BucketTitrator {
                     this->t = 0;
                     this->currentBucketIndex++;
 
-                    const double deltaZero = this->deltaZero;
+                    const float deltaZero = this->deltaZero;
                     int i = this->totalBuckets - 1 - this->currentBucketIndex + std::ceil( std::log(deltaZero) / std::log(1 + epsilon));
-                    const double threshold = std::pow(1 + epsilon, i);
+                    const float threshold = std::pow(1 + epsilon, i);
                     std::cout << "Bucket Threshold: " << threshold << std::endl;
                     //transfer current contents to next bucket
                     this->bucket = bucket->transferContents(threshold); //TODO:: change this to use the other constructor
@@ -129,13 +129,13 @@ class ThreeSieveBucketTitrator : public BucketTitrator {
         return this->firstBucketBuilt;
     }
 
-    void initBuckets(const double deltaZero) {
+    void initBuckets(const float deltaZero) {
         this->deltaZero = deltaZero;
         this->totalBuckets = this->getNumberOfBuckets(this->k, this->epsilon);
 
         std::cout << "number of buckets " << this->totalBuckets << " with deltaZero of " << deltaZero << std::endl;
         int i = this->totalBuckets - 1 + std::ceil( std::log(deltaZero) / std::log(1 + epsilon));
-        double threshold = std::pow(1 + epsilon, i);
+        float threshold = std::pow(1 + epsilon, i);
         std::cout << "Bucket Threshold: " << threshold << std::endl;
         this->bucket = std::make_unique<ThresholdBucket>(threshold, k);
         this->firstBucketBuilt = true;
@@ -146,7 +146,7 @@ class ThreeSieveBucketTitrator : public BucketTitrator {
 
 class SieveStreamingBucketTitrator : public BucketTitrator {
     private:
-    const double epsilon;
+    const float epsilon;
     const unsigned int numThreads;
     const unsigned int k;
     double deltaZero;
@@ -156,7 +156,7 @@ class SieveStreamingBucketTitrator : public BucketTitrator {
     public:
     SieveStreamingBucketTitrator(
         const unsigned int numThreads,
-        const double epsilon,
+        const float epsilon,
         const unsigned int k
     ) :  
         numThreads(numThreads),
@@ -228,10 +228,10 @@ class SieveStreamingBucketTitrator : public BucketTitrator {
     }
 
     std::unique_ptr<Subset> getBestSolutionDestroyTitrator() {
-        double bestBucketScore = 0;
+        float bestBucketScore = 0;
         size_t bestBucketIndex = -1;
         for (size_t i = 0; i < this->buckets.size(); i++) {
-            double bucketScore = this->buckets[i].getUtility();
+            float bucketScore = this->buckets[i].getUtility();
             if (bucketScore > bestBucketScore) {
                 bestBucketScore = bucketScore;
                 bestBucketIndex = i;
@@ -245,7 +245,7 @@ class SieveStreamingBucketTitrator : public BucketTitrator {
         return buckets.size() != 0;
     }
 
-    void initBuckets(const double deltaZero) {
+    void initBuckets(const float deltaZero) {
         const size_t numBuckets = this->getNumberOfBuckets(this->k, this->epsilon);
         this->deltaZero = deltaZero;
         std::cout << "number of buckets " << numBuckets << " with deltaZero of " << deltaZero << std::endl;
@@ -253,7 +253,7 @@ class SieveStreamingBucketTitrator : public BucketTitrator {
         for (int bucket = 0; bucket < numBuckets; bucket++)
         {
             int i = bucket + std::ceil( std::log(deltaZero) / std::log(1 + epsilon));
-            double threshold = (double)std::pow(1 + epsilon, i);
+            float threshold = (float)std::pow(1 + epsilon, i);
             this->buckets.push_back(ThresholdBucket(threshold, k));
         }
     }
